update .net version and cake refs
